+++ enviroment.js
@@ -1,3 +1,4 @@
 export const BASE_URL = 'https://kitsu.io/api/edge/';
+export const API_KEY_YOUTUBE = 'AIzaSyCo1SPfNRFu7aJUeBlv3E_1vJKorikp4HE';
 export const WIDTH_DESING = 375;
 export const HEIHGT_DESING = 667;