+++ src/components/custom_headers/styles.js
@@ -41,12 +41,6 @@ export default StyleSheet.create({
     paddingRight: 10,
   },
   // detail header
-  detailContainer: {
-    backgroundColor: colors.white,
-  },
-  titleContainer: {
-    justifyContent: 'center',
-  },
   title: {
     color: colors.titleColorText,
     fontSize: resize(16),