+++ src/screens/app/DetailData/index.js
@@ -1,13 +1,21 @@
 import React, {Component} from 'react';
-import {Text} from 'react-native';
-import {View, Container, Thumbnail, Content} from 'native-base';
+import {Text, FlatList} from 'react-native';
+import {View, Container, Content} from 'native-base';
+import moment from 'moment';
+
+// custom
 import Header from '../../../components/header';
 import ImageNotFound from '../../../assets/img/img_not_found.png';
-
+import StarRating from '../../../components/star_rating';
+import PosterImage from '../../../components/poster_image';
+import YearReleasedAndFinshied from '../../../components/year_released_finished';
+import AiringStatus from '../../../components/airing_status';
+import LoadingAnimation from '../../../components/loading_animation';
 import styles from './styles';
-import moment from 'moment';
+
+// api
 import {getGenresForResource} from '../../../api/kitsu';
-import {FlatList} from 'react-native-gesture-handler';
+import {ScrollView} from 'react-native-gesture-handler';
 
 class DetailData extends Component {
   constructor(props) {
@@ -29,13 +37,12 @@ class DetailData extends Component {
         name: genre.attributes.name,
       };
     });
-    console.log(genres);
+    console.log(detail);
     this.setState({
       title: detail.attributes.canonicalTitle,
       genres,
       isLoading: false,
     });
-    console.log(detail);
   };
 
   validateImage = image => {
@@ -47,7 +54,6 @@ class DetailData extends Component {
   };
 
   getDataTitle = title => {
-    console.log(title);
     if (title && Object.keys(title).length) {
       const keys = Object.keys(title);
       const key = keys.filter(k => {
@@ -55,8 +61,6 @@ class DetailData extends Component {
           return title;
         }
       })[0];
-      console.log(key);
-      // const key = Object.keys(title)[0];
       return title[key];
     }
     return 'Nel';
@@ -71,97 +75,76 @@ class DetailData extends Component {
     }
   };
 
+  validateStringEpisodes = episodes => {
+    if (episodes) {
+      const episodesText = episodes == 1 ? 'Episode' : 'Episodes';
+      return `${episodes} ${episodesText}`;
+    } else {
+      return '0 Episodes';
+    }
+  };
+
   render() {
     const {title, genres, isLoading} = this.state;
     const detail = this.props.navigation.getParam('detail', {});
+    const {
+      titles,
+      startDate,
+      endDate,
+      status,
+      ageRating,
+      ageRatingGuide,
+      episodeCount,
+      posterImage,
+      youtubeVideoId,
+      userCount,
+      synopsis,
+      averageRating,
+    } = detail.attributes;
     return (
       <Container>
         <Header title={title} type="detail" />
         {isLoading ? (
-          <Text>Cargando</Text>
+          <LoadingAnimation type="detail" />
         ) : (
-          <View>
-            <View
-              style={{
-                paddingVertical: 20,
-                paddingLeft: 10,
-                flexDirection: 'row',
-              }}>
-              <View style={{flex: 1}}>
-                <Thumbnail
-                  square
-                  style={styles.image}
-                  resizeMode="stretch"
-                  source={this.validateImage(detail.attributes.posterImage)}
-                />
-              </View>
-              <View style={{flex: 1}}>
-                <Text style={styles.titleHeader}>Main Title</Text>
-                <Text style={styles.titleInfo}>
-                  {this.getDataTitle(detail.attributes.titles)}
-                </Text>
-
-                <Text style={[styles.titleHeader, {paddingTop: 10}]}>
-                  Canonical Title
-                </Text>
-                <Text style={styles.titleInfo}>
-                  {detail.attributes.canonicalTitle}
-                </Text>
-
-                <Text style={[styles.titleHeader, {paddingTop: 10}]}>Type</Text>
-                <View style={{flexDirection: 'row'}}>
-                  <Text style={styles.titleType}>{detail.type}</Text>
-                  <Text style={styles.titleType}>
-                    {`${detail.attributes.episodeLength} Eps`}
-                  </Text>
-                </View>
-
-                <Text style={[styles.titleHeader, {paddingTop: 10}]}>Year</Text>
-                <Text style={styles.titleInfo}>
-                  {this.showYear(detail.attributes)}
-                </Text>
-              </View>
+          <Content style={styles.content}>
+            <Text style={styles.titleHeader}>{this.getDataTitle(titles)}</Text>
+            <YearReleasedAndFinshied
+              releasedYear={startDate}
+              finishedYear={endDate}
+            />
+            <ScrollView horizontal showsHorizontalScrollIndicator={false}>
+              <AiringStatus status={status} />
+              <Text style={styles.ageRating}>
+                {ageRatingGuide || ageRating}
+              </Text>
+              <Text style={styles.episodes}>
+                {this.validateStringEpisodes(episodeCount)}
+              </Text>
+            </ScrollView>
+            <PosterImage
+              image={this.validateImage(posterImage)}
+              videoUrl={youtubeVideoId}
+            />
+            <View style={styles.contentGenres}>
+              <FlatList
+                data={genres}
+                horizontal
+                showsHorizontalScrollIndicator={false}
+                renderItem={({item}) => {
+                  return (
+                    <Text style={styles.genres} key={item.id}>
+                      {item.name}
+                    </Text>
+                  );
+                }}
+              />
             </View>
-            <View style={{padding: 10}}>
-              <Text style={styles.titleHeader}>Genres</Text>
-              <View style={{flexDirection: 'row'}}>
-                <FlatList
-                  data={genres}
-                  horizontal
-                  showsHorizontalScrollIndicator={false}
-                  renderItem={({item}) => {
-                    return (
-                      <Text style={styles.genres} key={item.id}>
-                        {item.name}
-                      </Text>
-                    );
-                  }}
-                />
-              </View>
-              <View style={{paddingTop: 10, flexDirection: 'row'}}>
-                <View style={{flexDirection: 'column', flex: 1}}>
-                  <Text style={styles.titleHeader}>Average Rating</Text>
-                  <Text style={styles.titleInfo}>
-                    {detail.attributes.averageRating}
-                  </Text>
-                </View>
-                <View style={{flexDirection: 'column', flex: 1}}>
-                  <Text style={styles.titleHeader}>Age Rating</Text>
-                  <Text style={styles.titleInfo}>
-                    {detail.attributes.ageRatingGuide}
-                  </Text>
-                </View>
-              </View>
-              <View style={{paddingTop: 10, paddingBottom: 10}}>
-                <Text style={styles.titleHeader}>Sipnosis</Text>
-                <Content style={{height: 90, width: '100%'}}>
-                  <Text style={styles.titleInfo}>
-                    {detail.attributes.synopsis}
-                  </Text>
-                </Content>
-              </View>
+            <StarRating rating={averageRating} userCount={userCount} />
+            <View style={styles.synopsis}>
+              <Text>{synopsis}</Text>
             </View>
-          </View>
+          </Content>
         )}
       </Container>
     );