+++ src/screens/app/DetailData/styles.js
@@ -3,18 +3,42 @@ import {resize} from '../../../utils/styles';
 import colors from '../../../utils/colors';
 
 export default StyleSheet.create({
+  content: {
+    paddingHorizontal: resize(15),
+  },
+  contentGenres: {flexDirection: 'row', paddingTop: 10, paddingBottom: 10},
+  synopsis: {paddingVertical: resize(10, 'h')},
   image: {
     borderRadius: resize(5),
     width: resize(170),
     height: resize(260, 'h'),
   },
   titleHeader: {
-    fontSize: resize(16),
+    fontSize: resize(20),
+    textAlign: 'left',
     color: colors.titleColorText,
   },
   titleInfo: {
     fontSize: resize(13),
     color: colors.colorText,
+  },
+  ageRating: {
+    marginLeft: resize(5),
+    marginTop: resize(10, 'h'),
+    backgroundColor: '#212121',
+    fontSize: resize(13),
+    paddingHorizontal: resize(5),
+    borderRadius: resize(10),
+    color: colors.white,
+  },
+  episodes: {
+    marginLeft: resize(5),
+    marginTop: resize(10, 'h'),
+    backgroundColor: '#00701a',
+    fontSize: resize(13),
+    paddingHorizontal: resize(5),
+    borderRadius: resize(10),
+    color: colors.white,
   },
   titleType: {
     borderColor: colors.primaryLight,
@@ -38,6 +62,6 @@ export default StyleSheet.create({
     paddingVertical: resize(3),
     paddingHorizontal: resize(10),
     fontWeight: 'bold',
-    backgroundColor: colors.primary,
+    backgroundColor: '#0277bd',
   },
 });